+++ CloudShell.Providers.Applications/ApplicationSettingReferenceDisplay.cs
@@ -1,6 +1,5 @@
 using CloudShell.Abstractions.Authorization;
 using CloudShell.Abstractions.ResourceManager;
-using CloudShell.Components;
 
 namespace CloudShell.Providers.Applications;